requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
Pillow==10.2.0
numpy==1.24.3
opencv-python==4.8.0.76
tensorflow==2.12.0
keras==2.12.0
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2
